requirements.txt
flask
langchain
langchain_groq
pinecone-client
langchain-pinecone
huggingface-hub
langchain-community
langchain-huggingface
